.history/requirements_20250813171018.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
SQLAlchemy==2.0.31
psycopg2-binary==2.9.9
python-dotenv==1.0.1
PyJWT==2.9.0
